requirements.txt
selenium==3.141.0
python-dotenv==0.15.0
pause==0.3